add shader base
