+++ src/compute_wrapper_impl.h
@@ -6,11 +6,12 @@
 #include "dawn_utils.hpp"
 #include "dawn_wrapper.h"
 #include "encoder_wrapper_impl.h"
+#include "shader_base.hpp"
 
 using namespace wgpu;
 
 namespace dawn_wrapper {
-struct compute_wrapper::pimpl {
+struct compute_wrapper::pimpl : private shader_base {
     pimpl() = delete;
 
     pimpl(Device device)
@@ -25,7 +26,7 @@ struct compute_wrapper::pimpl {
     void compile_shader(std::string script, std::string entryPoint)
     {
         m_shader = dawn_utils::make_compute_shader(m_device, script, entryPoint.c_str());
-        m_shader.GetCompilationInfo(&compilation_callback, this);
+        m_shader.GetCompilationInfo(&shader_base::compilation_callback, this);
         m_entryPoint = entryPoint;
     }
 
@@ -68,37 +69,6 @@ struct compute_wrapper::pimpl {
     bindgroup_wrapper make_bindgroup()
     {
         return std::make_shared<bindgroup_wrapper::pimpl>(m_entryPoint);
-    }
-
-private:
-    static void compilation_callback(WGPUCompilationInfoRequestStatus status, struct WGPUCompilationInfo const* compilationInfo, void* userdata)
-    {
-        pimpl* instance(reinterpret_cast<pimpl*>(userdata));
-        std::stringstream messages;
-        size_t errorCount = 0;
-        for (auto i = 0; i < compilationInfo->messageCount; ++i) {
-            const auto message = compilationInfo->messages[i];
-            if (message.type == WGPUCompilationMessageType_Error) {
-                messages << "Error(" << i << "): ";
-                ++errorCount;
-            } else if (message.type == WGPUCompilationMessageType_Warning) {
-                messages << "Warning(" << i << "): ";
-            } else if (message.type == WGPUCompilationMessageType_Info) {
-                messages << "Info(" << i << "): ";
-            }
-#ifndef __EMSCRIPTEN__
-            messages << message.message.data << std::endl;
-#else
-            messages << message.message << std::endl;
-#endif
-        }
-
-        std::cout << messages.str() << std::endl;
-        //      instance->m_shaderCompileCallback(messages.str());
-
-        //        if (!errorCount) {
-        //            instance->setFragmentShaderUser();
-        //        }
     }
 
     Device m_device;
